Use XML escaping rather than HTML escaping.
There is no HTML involved in this app, only XML. The .escapeHtml method
produces entities like &eacute; for non-ASCII characters which the subsequent XML parser doesn't
recognize. Since this ends up as text in an XML element, technically
only <, >, and & are required to be escaped.

Obviates the need for passing `_has_escapes` and escaping yourself to
work around this to support non-ASCII characters, like in #21.